# laravel-with-jwt/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get posts list",
                "operationId": "3fda362e2e082b5357ac416e7767e435",
                "responses": {
                    "200": {
                        "description": "Success, return data list"
                    },
                    "500": {
                        "description": "Error, return error message"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create new posts",
                "operationId": "dcb66200c3132996151884ac232cf439",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "content": {
                                        "description": "Content",
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "400": {
                        "description": "Error, return error list"
                    },
                    "500": {
                        "description": "Error, return error message"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get post by ID",
                "operationId": "142b5adb0520cca80cedfe09d9298332",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, return data"
                    },
                    "404": {
                        "description": "Data not found"
                    },
                    "500": {
                        "description": "Error, return error message"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update post by ID",
                "operationId": "650f4bed0846d9350903d8a17c7b503a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "content": {
                                        "description": "Content",
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "400": {
                        "description": "Error, return error list"
                    },
                    "404": {
                        "description": "Data not found"
                    },
                    "500": {
                        "description": "Error, return error message"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete post by ID",
                "operationId": "9d99f2e42af2f722a058a86245aa856f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data deleted"
                    },
                    "404": {
                        "description": "Data not found"
                    },
                    "500": {
                        "description": "Error, return error message"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}